requirements.txt
fastapi
uvicorn
requests
pynacl